requirements.txt
networkx
pyYAML